+++ modules/alloy/src/main/ts/ephox/alloy/positioning/layout/Layout.ts
@@ -2,6 +2,7 @@ import { nu as NuSpotInfo } from '../view/SpotInfo';
 import { Bubble } from './Bubble';
 import * as Direction from './Direction';
 import { boundsRestriction, AnchorBoxBounds } from './LayoutBounds';
+import * as LayoutLabels from './LayoutLabels';
 import { AnchorBox, AnchorElement, AnchorLayout } from './LayoutTypes';
 
 /*
@@ -38,7 +39,9 @@ const southeast: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubb
   bubbles.southeast(),
   Direction.southeast(),
   boundsRestriction(anchor, { left: AnchorBoxBounds.LeftEdge, top: AnchorBoxBounds.BottomEdge }),
-  'layout-se');
+  LayoutLabels.southEast,
+  false
+);
 
 const southwest: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
   westX(anchor, element),
@@ -46,7 +49,8 @@ const southwest: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubb
   bubbles.southwest(),
   Direction.southwest(),
   boundsRestriction(anchor, { right: AnchorBoxBounds.RightEdge, top: AnchorBoxBounds.BottomEdge }),
-  'layout-sw'
+  LayoutLabels.southWest,
+  false
 );
 
 const northeast: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -55,7 +59,8 @@ const northeast: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubb
   bubbles.northeast(),
   Direction.northeast(),
   boundsRestriction(anchor, { left: AnchorBoxBounds.LeftEdge, bottom: AnchorBoxBounds.TopEdge }),
-  'layout-ne'
+  LayoutLabels.northEast,
+  false
 );
 
 const northwest: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -64,7 +69,8 @@ const northwest: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubb
   bubbles.northwest(),
   Direction.northwest(),
   boundsRestriction(anchor, { right: AnchorBoxBounds.RightEdge, bottom: AnchorBoxBounds.TopEdge }),
-  'layout-nw'
+  LayoutLabels.northWest,
+  false
 );
 
 const north: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -73,7 +79,8 @@ const north: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles:
   bubbles.north(),
   Direction.north(),
   boundsRestriction(anchor, { bottom: AnchorBoxBounds.TopEdge }),
-  'layout-n'
+  LayoutLabels.north,
+  false
 );
 
 const south: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -82,7 +89,8 @@ const south: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles:
   bubbles.south(),
   Direction.south(),
   boundsRestriction(anchor, { top: AnchorBoxBounds.BottomEdge }),
-  'layout-s'
+  LayoutLabels.south,
+  false
 );
 
 const east: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -91,7 +99,8 @@ const east: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: 
   bubbles.east(),
   Direction.east(),
   boundsRestriction(anchor, { left: AnchorBoxBounds.RightEdge }),
-  'layout-e'
+  LayoutLabels.east,
+  false
 );
 
 const west: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: Bubble) => NuSpotInfo(
@@ -100,7 +109,8 @@ const west: AnchorLayout = (anchor: AnchorBox, element: AnchorElement, bubbles: 
   bubbles.west(),
   Direction.west(),
   boundsRestriction(anchor, { right: AnchorBoxBounds.LeftEdge }),
-  'layout-w'
+  LayoutLabels.west,
+  false
 );
 
 const all = (): AnchorLayout[] => [ southeast, southwest, northeast, northwest, south, north, east, west ];