+++ modules/alloy/src/main/ts/ephox/alloy/positioning/layout/LayoutLabels.ts
@@ -1,0 +1,66 @@
+const southEast = 'tox-layout-se';
+const southWest = 'tox-layout-sw';
+const northEast = 'tox-layout-ne';
+const northWest = 'tox-layout-nw';
+const north = 'tox-layout-n';
+const south = 'tox-layout-s';
+const east = 'tox-layout-e';
+const west = 'tox-layout-w';
+
+const southEastInner = 'tox-layout-inner-se';
+const southWestInner = 'tox-layout-inner-sw';
+const northEastInner = 'tox-layout-inner-ne';
+const northWestInner = 'tox-layout-inner-nw';
+const northInner = 'tox-layout-inner-n';
+const southInner = 'tox-layout-inner-s';
+const eastInner = 'tox-layout-inner-e';
+const westInner = 'tox-layout-inner-w';
+
+const northPinned = 'tox-layout-pinned-n';
+const southPinned = 'tox-layout-pinned-s';
+
+const contextBarTransitionClass = 'tox-context-bar-layout-transition-animation';
+
+const all = [
+  southEast,
+  northEast,
+  southWest,
+  northWest,
+  south,
+  north,
+  east,
+  west,
+  southEastInner,
+  northEastInner,
+  southWestInner,
+  northWestInner,
+  southInner,
+  northInner,
+  eastInner,
+  westInner,
+  northPinned,
+  southPinned
+];
+
+export {
+  all,
+  southEast,
+  northEast,
+  southWest,
+  northWest,
+  south,
+  north,
+  east,
+  west,
+  southEastInner,
+  northEastInner,
+  southWestInner,
+  northWestInner,
+  southInner,
+  northInner,
+  eastInner,
+  westInner,
+  northPinned,
+  southPinned,
+  contextBarTransitionClass
+};