+++ modules/tinymce/src/themes/silver/main/ts/ui/context/ContextUi.ts
@@ -6,7 +6,7 @@
  */
 
 import {
-  AddEventsBehaviour, AlloyComponent, AlloyEvents, AlloySpec, AlloyTriggers, Behaviour, CustomEvent, GuiFactory, InlineView, Keying, NativeEvents
+  AddEventsBehaviour, AlloyComponent, AlloyEvents, AlloySpec, AlloyTriggers, Behaviour, contextBarTransitionClass, CustomEvent, GuiFactory, InlineView, Keying, NativeEvents
 } from '@ephox/alloy';
 import { Arr, Cell, Id, Optional, Result } from '@ephox/katamari';
 import { Class, Css, EventArgs, Focus, SugarElement, SugarShadowDom, Width } from '@ephox/sugar';
@@ -51,9 +51,13 @@ const renderContextToolbar = (spec: { onEscape: () => Optional<boolean>; sink: A
 
     inlineBehaviours: Behaviour.derive([
       AddEventsBehaviour.config('context-toolbar-events', [
-        AlloyEvents.runOnSource<EventArgs>(NativeEvents.transitionend(), (comp, _se) => {
-          Class.remove(comp.element, resizingClass);
-          Css.remove(comp.element, 'width');
+        AlloyEvents.runOnSource<EventArgs<TransitionEvent>>(NativeEvents.transitionend(), (comp, se) => {
+          if (se.event.raw.propertyName === 'top' || se.event.raw.propertyName === 'bottom') {
+            Class.remove(comp.element, contextBarTransitionClass);
+          } else {
+            Class.remove(comp.element, resizingClass);
+            Css.remove(comp.element, 'width');
+          }
         }),
 
         AlloyEvents.run<ChangeSlideEvent>(changeSlideEvent, (comp, se) => {